vmcheck: adapt for non-vagrant hosts
This is just the final bit required to make sure the vagrant and
non-vagrant paths can work happily together. It's mostly minor fixes,
though the most major change which also affects vagrant is that we now
sync to the root home dir, rather than ~vagrant.

+++ tests/vmcheck/sync.sh
@@ -1,10 +1,32 @@
 #!/bin/bash
 set -euo pipefail
 
-. ${commondir}/libvm.sh
+if test -z "${INSIDE_VM:-}"; then
 
-set -x
-cd ${builddir}
-rm insttree -rf
-make install DESTDIR=$(pwd)/insttree
-ssh -o User=root -F ssh-config vmcheck "ostree admin unlock; cd ~vagrant/sync && rsync -rv insttree/usr/ /usr/ && systemctl restart rpm-ostreed"
+    # do this in the host
+    . ${commondir}/libvm.sh
+    vm_setup
+
+    if ! vm_ssh_wait 30; then
+      echo "ERROR: A running VM is required for 'make vmcheck'."
+      exit 1
+    fi
+
+    set -x
+
+    topdir=$(git rev-parse --show-toplevel)
+    cd ${topdir}
+    rm insttree -rf
+    make install DESTDIR=$(pwd)/insttree
+    vm_rsync
+
+    $SSH "sudo env INSIDE_VM=1 /var/roothome/sync/tests/vmcheck/sync.sh"
+    exit 0
+else
+
+    # then do this in the VM
+    set -x
+    ostree admin unlock || :
+    rsync -rv /var/roothome/sync/insttree/usr/ /usr/
+    systemctl restart rpm-ostreed
+fi
